[feat/#52]: 컴포넌트 분리
